refactor: move new intelligentSearch hook to ila26
